Move crc32calc.py to utils
