+++ sabnzbd/utils/crc32calc.py
